pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ai-digest"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = ["feedparser", "pyyaml"]

[project.scripts]
ai-digest = "ai_digest.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_digest*"]
